remove existing detail view table
