--- client/packages/invoices/src/OutboundShipment/DetailView/ContentArea.tsx
@@ -1,124 +1,0 @@
-import React, { FC, useEffect } from 'react';
-import {
-  DataTable,
-  useTranslation,
-  useIsGrouped,
-  InvoiceLineNodeType,
-  useRowStyle,
-  AppSxProp,
-  NothingHere,
-  useUrlQueryParams,
-  useFeatureFlags,
-  MaterialTable,
-} from '@openmsupply-client/common';
-import { MRT_TableInstance } from 'material-react-table';
-import { useOutbound } from '../api';
-import { useOutboundColumns } from './columns';
-import { StockOutLineFragment } from '../../StockOut';
-import { StockOutItem } from '../../types';
-import { Expand } from './ExpandoTable';
-
-interface ContentAreaProps {
-  onAddItem: () => void;
-  onRowClick?: null | ((rowData: StockOutLineFragment | StockOutItem) => void);
-  table?: MRT_TableInstance<StockOutLineFragment>;
-  // onReturnLines: (selectedLines: StockOutLineFragment[]) => void;
-}
-
-const useHighlightPlaceholderRows = (
-  rows: StockOutLineFragment[] | StockOutItem[] | undefined
-) => {
-  const { setRowStyles } = useRowStyle();
-
-  useEffect(() => {
-    if (!rows) return;
-    const placeholders = [];
-
-    // This is a verbose .filter() on `rows` to find the placeholder lines.
-    // There is an issue with using `filter()` on a type which is
-    // A[] | B[]
-    // https://github.com/microsoft/TypeScript/issues/44373
-    for (const row of rows) {
-      if ('type' in row) {
-        if (
-          row.type === InvoiceLineNodeType.UnallocatedStock ||
-          row.numberOfPacks === 0
-        ) {
-          placeholders.push(row.id);
-        }
-      } else {
-        const hasPlaceholder = row.lines.some(
-          line => line.type === InvoiceLineNodeType.UnallocatedStock
-        );
-        if (hasPlaceholder) {
-          // Add both the OutboundItem and the individual lines, as
-          // this will cause the item to be highlighted as well as the
-          // lines within the expansion when grouped.
-          row.lines.forEach(line => {
-            if (line.type === InvoiceLineNodeType.UnallocatedStock) {
-              placeholders.push(line.id);
-            }
-          });
-          placeholders.push(row.id);
-        }
-      }
-    }
-
-    const style: AppSxProp = {
-      color: theme => theme.palette.secondary.light,
-    };
-    setRowStyles(placeholders, style);
-  }, [rows, setRowStyles]);
-};
-
-export const ContentAreaComponent: FC<ContentAreaProps> = ({
-  onAddItem,
-  onRowClick,
-  // onReturnLines,
-  table,
-}) => {
-  const t = useTranslation();
-
-  const {
-    updateSortQuery,
-    queryParams: { sortBy },
-  } = useUrlQueryParams();
-  const { isGrouped } = useIsGrouped('outboundShipment');
-  const { rows } = useOutbound.line.rows(isGrouped);
-  const columns = useOutboundColumns({
-    onChangeSortBy: updateSortQuery,
-    sortBy,
-  });
-  const isDisabled = useOutbound.utils.isDisabled();
-  useHighlightPlaceholderRows(rows);
-  const { tableUsabilityImprovements } = useFeatureFlags();
-
-  if (!rows) return null;
-
-  return (
-    <>
-      {tableUsabilityImprovements && table ? (
-        <MaterialTable table={table} />
-      ) : (
-        <DataTable
-          id="outbound-detail"
-          onRowClick={onRowClick}
-          ExpandContent={props => <Expand {...props} />}
-          columns={columns}
-          data={rows}
-          enableColumnSelection
-          noDataElement={
-            <NothingHere
-              body={t('error.no-outbound-items')}
-              onCreate={isDisabled ? undefined : () => onAddItem()}
-              buttonText={t('button.add-item')}
-            />
-          }
-          isRowAnimated={true}
-        />
-      )}
-    </>
-  );
-};
-
-export const ContentArea = React.memo(ContentAreaComponent);
