+++ client/packages/invoices/src/OutboundShipment/DetailView/DetailView.tsx
@@ -1,30 +1,22 @@
-import React, { useCallback, useEffect, useMemo } from 'react';
+import React, { useCallback, useEffect } from 'react';
 import {
-  TableProvider,
-  createTableStore,
   useEditModal,
   DetailViewSkeleton,
   AlertModal,
   useNavigate,
   RouteBuilder,
   useTranslation,
-  createQueryParamsStore,
   DetailTabs,
   ModalMode,
   useNotification,
-  useTableStore,
   useBreadcrumbs,
   useNonPaginatedMaterialTable,
-  usePreferences,
   InvoiceLineNodeType,
   useIsGrouped,
-  ColumnDef,
-  ArrayUtils,
-  Groupable,
-  ColumnType,
+  MaterialTable,
+  NothingHere,
 } from '@openmsupply-client/common';
 import { toItemRow, ActivityLogList } from '@openmsupply-client/system';
-import { ContentArea } from './ContentArea';
 import { StockOutItem } from '../../types';
 import { Toolbar } from './Toolbar';
 import { Footer } from './Footer';
@@ -37,8 +29,9 @@ import { CustomerReturnEditModal } from '../../Returns';
 import { canReturnOutboundLines } from '../../utils';
 import { OutboundLineEdit, OutboundOpenedWith } from './OutboundLineEdit';
 import { useOutboundLines } from '../api/hooks/line/useOutboundLines';
+import { useOutboundColumns } from './columns';
 
-const DetailViewInner = () => {
+export const DetailView = () => {
   const t = useTranslation();
   const { info } = useNotification();
   const isDisabled = useOutbound.utils.isDisabled();
@@ -53,7 +46,6 @@ const DetailViewInner = () => {
     mode: returnModalMode,
     setMode: setReturnMode,
   } = useEditModal<string[]>();
-  const { manageVvmStatusForStock } = usePreferences();
 
   const { data, isLoading } = useOutbound.document.get();
   const { isGrouped } = useIsGrouped('outboundShipment');
@@ -71,7 +63,6 @@ const DetailViewInner = () => {
     onOpen(openWith);
     setMode(ModalMode.Create);
   };
-  const { clearSelected } = useTableStore();
 
   const onReturn = async (selectedLines: StockOutLineFragment[]) => {
     if (!data || !canReturnOutboundLines(data)) {
@@ -92,181 +83,33 @@ const DetailViewInner = () => {
     setCustomBreadcrumbs({ 1: data?.invoiceNumber.toString() ?? '' });
   }, [setCustomBreadcrumbs, data?.invoiceNumber]);
 
-  const isDefaultPlaceholderRow = (row: StockOutLineFragment) =>
-    row.type === InvoiceLineNodeType.UnallocatedStock && !row.numberOfPacks;
+  const columns = useOutboundColumns();
 
-  const mrtColumns = useMemo(() => {
-    const cols: ColumnDef<Groupable<StockOutLineFragment>>[] = [
-      {
-        accessorKey: 'item.code',
-        header: t('label.code'),
-        size: 120,
-        pin: 'left',
-        enableColumnFilter: true,
-        enableSorting: true,
-      },
-      {
-        accessorKey: 'itemName',
-        header: t('label.name'),
-        size: 400,
-        enableColumnFilter: true,
-        enableSorting: true,
-      },
-      {
-        accessorKey: 'batch',
-        header: t('label.batch'),
-        defaultHideOnMobile: true,
-      },
-      {
-        id: 'expiryDate',
-        // expiryDate is a string - use accessorFn to convert to Date object for sort and filtering
-        accessorFn: row => (row.expiryDate ? new Date(row.expiryDate) : null),
-        header: t('label.expiry-date'),
-        columnType: ColumnType.Date,
-        defaultHideOnMobile: true,
-        enableColumnFilter: true,
-        enableSorting: true,
-      },
-      {
-        id: 'vvmStatus',
-        accessorFn: row => row.vvmStatus?.description ?? '',
-        header: t('label.vvm-status'),
-        includeColumn: manageVvmStatusForStock,
-        defaultHideOnMobile: true,
-      },
-      {
-        id: 'locationCode',
-        accessorFn: row => row.location?.code ?? '',
-        header: t('label.location'),
-        enableColumnFilter: true,
-        defaultHideOnMobile: true,
-      },
-      {
-        id: 'itemUnit',
-        accessorKey: 'item.unitName',
-        header: t('label.unit-name'),
-        filterVariant: 'select',
-        defaultHideOnMobile: true,
-      },
-      {
-        accessorKey: 'packSize',
-        header: t('label.pack-size'),
-        columnType: ColumnType.Number,
-        defaultHideOnMobile: true,
-      },
-      {
-        id: 'itemDoses',
-        header: t('label.doses-per-unit'),
-        columnType: ColumnType.Number,
-        defaultHideOnMobile: true,
-        accessorFn: row => (row.item.isVaccine ? row.item.doses : undefined),
-      },
-      {
-        accessorKey: 'numberOfPacks',
-        header: t('label.pack-quantity'),
-        columnType: ColumnType.Number,
-        accessorFn: row => {
-          if ('subRows' in row)
-            return ArrayUtils.getSum(row.subRows ?? [], 'numberOfPacks');
-
-          return row.numberOfPacks;
-        },
-      },
-      {
-        id: 'unitQuantity',
-        header: t('label.unit-quantity'),
-        description: t('description.unit-quantity'),
-        columnType: ColumnType.Number,
-        defaultHideOnMobile: true,
-        accessorFn: row => {
-          if ('subRows' in row)
-            return ArrayUtils.getUnitQuantity(row.subRows ?? []);
-
-          return row.packSize * row.numberOfPacks;
-        },
-      },
-      {
-        id: 'doseQuantity',
-        header: t('label.doses'),
-        columnType: ColumnType.Number,
-        defaultHideOnMobile: true,
-        accessorFn: row => {
-          if (!row.item.isVaccine) return null;
-          if ('subRows' in row)
-            return (
-              ArrayUtils.getUnitQuantity(row.subRows ?? []) *
-              (row.item.doses ?? 1)
-            );
-
-          return row.packSize * row.numberOfPacks * (row.item.doses ?? 1);
-        },
-      },
-      {
-        id: 'unitSellPrice',
-        header: t('label.unit-sell-price'),
-        columnType: ColumnType.Currency,
-        defaultHideOnMobile: true,
-        accessorFn: rowData => {
-          if ('subRows' in rowData) {
-            return ArrayUtils.getAveragePrice(
-              rowData.subRows ?? [],
-              'sellPricePerPack'
-            );
-          } else {
-            if (isDefaultPlaceholderRow(rowData)) return undefined;
-            return (rowData.sellPricePerPack ?? 0) / rowData.packSize;
-          }
-        },
-      },
-      {
-        id: 'total',
-        header: t('label.total'),
-        columnType: ColumnType.Currency,
-        defaultHideOnMobile: true,
-        accessorFn: rowData => {
-          if ('subRows' in rowData) {
-            return Object.values(rowData.subRows ?? []).reduce(
-              (sum, batch) =>
-                sum + batch.sellPricePerPack * batch.numberOfPacks,
-              0
-            );
-          } else {
-            if (isDefaultPlaceholderRow(rowData)) return '';
-
-            const x = rowData.sellPricePerPack * rowData.numberOfPacks;
-            return x;
-          }
-        },
-      },
-    ];
-
-    return cols;
-  }, [manageVvmStatusForStock]);
-
-  const { table, selectedRows, resetRowSelection } =
+  const { table, selectedRows } =
     useNonPaginatedMaterialTable<StockOutLineFragment>({
       tableId: 'outbound-shipment-detail-view',
-      columns: mrtColumns,
+      columns,
       data: rows ?? [],
-      onRowClick: onRowClick ? row => onRowClick(row) : () => {},
+      groupByField: isGrouped ? 'itemName' : undefined,
       isLoading: false,
+      onRowClick: isDisabled ? onRowClick : undefined,
       getIsPlaceholderRow: row =>
         row.type === InvoiceLineNodeType.UnallocatedStock ||
         row.numberOfPacks === 0,
-      groupByField: isGrouped ? 'itemName' : undefined,
+      renderEmptyRowsFallback: () => (
+        <NothingHere
+          body={t('error.no-outbound-items')}
+          onCreate={isDisabled ? undefined : () => onAddItem()}
+          buttonText={t('button.add-item')}
+        />
+      ),
     });
 
   if (isLoading) return <DetailViewSkeleton hasGroupBy={true} hasHold={true} />;
 
   const tabs = [
     {
-      Component: (
-        <ContentArea
-          onRowClick={!isDisabled ? onRowClick : null}
-          onAddItem={onAddItem}
-          table={table}
-        />
-      ),
+      Component: <MaterialTable table={table} />,
       value: 'Details',
     },
     {
@@ -301,7 +144,7 @@ const DetailViewInner = () => {
               customerId={data.otherPartyId}
               modalMode={returnModalMode}
               outboundShipmentId={data.id}
-              onCreate={clearSelected}
+              onCreate={table.resetRowSelection}
               isNewReturn
             />
           )}
@@ -310,8 +153,8 @@ const DetailViewInner = () => {
           <DetailTabs tabs={tabs} />
           <Footer
             onReturnLines={onReturn}
-            selectedRows={selectedRows as StockOutLineFragment[]}
-            resetRowSelection={resetRowSelection}
+            selectedRows={selectedRows}
+            resetRowSelection={table.resetRowSelection}
           />
           <SidePanel />
         </>
@@ -330,22 +173,5 @@ const DetailViewInner = () => {
         />
       )}
     </React.Suspense>
-  );
-};
-
-export const DetailView = () => {
-  return (
-    <TableProvider
-      createStore={createTableStore}
-      queryParamsStore={createQueryParamsStore<
-        StockOutLineFragment | StockOutItem
-      >({
-        initialSortBy: {
-          key: 'itemName',
-        },
-      })}
-    >
-      <DetailViewInner />
-    </TableProvider>
   );
 };
